+++ CHANGELOG.md
@@ -4,3 +4,4 @@ v0.1.1 - XXXX-XX-XX
 
 Added
 - New dispatch option `atomic_priority_set` to set priority and dispatch event in one atomic operation.
+- New dispatch option for `members` called `external` to dispatch event to nodes except the local node.